+++ plugins/postcss-design-tokens/src/constants.ts
@@ -1,2 +1,4 @@
 // a random, but shared default condition
 export const DEFAULT_CONDITION = '6b4e71e7-4787-42f7-a092-8684961895db';
+export const DEFAULT_VALUE_FUNCTION_NAME = 'design-token';
+export const DEFAULT_IMPORT_AT_RULE_NAME = 'design-tokens';