+++ plugins/postcss-design-tokens/src/index.ts
@@ -23,7 +23,7 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 				Once: async (root, { result }) => {
 					const designTokenAtRules: Array<{filePath: string, params: string, node: Node}> = [];
 					root.walkAtRules((atRule) => {
-						if (atRule.name.toLowerCase() !== 'design-tokens') {
+						if (atRule.name.toLowerCase() !== options.importAtRuleName) {
 							return;
 						}
 
@@ -65,7 +65,7 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 					}
 				},
 				Declaration(decl, { result }) {
-					if (decl.value.toLowerCase().indexOf('design-token') === -1) {
+					if (!decl.value.toLowerCase().includes(options.valueFunctionName)) {
 						return;
 					}