+++ plugins/postcss-design-tokens/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes to PostCSS Design Tokens
 
+### Unreleased
+
+- Added `valueFunctionName` option to control the `design-token` function name.
+- Added `importAtRuleName` option to control the `design-tokens` at-rule name.
+
 ### 1.0.0 (May 23, 2022)
 
 - Initial version